Bug fix warning section deletion (#114)
* fix(): section deleteion restriction with test

* fix(): changed the functionality of section delete

* fix(): removed the deleted warning from the generic deletepop component and passedthe warning as prop

* chore(): synced with dev

+++ app/routes/sections.tsx
@@ -14,6 +14,7 @@ import {
   useSubmit,
 } from '@remix-run/react'
 import {
+  checkSectionById,
   createSection,
   deleteSectionById,
   getAllSections,
@@ -128,22 +129,43 @@ export const action: ActionFunction = async ({ request }) => {
     return addHandle
   }
   let deleteHandle = null
+  let isSectionDelete = false
+  let isTestDeleted: Array<boolean> | undefined
   if (action === 'sectionDelete') {
-    await deleteSectionById(formData.get('id') as string)
-      .then((res) => {
-        deleteHandle = json<ActionData>(
-          { resp: { status: statusCheck.deletedSuccess } },
-
-          { status: 200 }
-        )
-      })
-      .catch((err) => {
-        let title = statusCheck.commonError
+    await checkSectionById(formData.get('id') as string).then((res) => {
+      if (res?._count.sectionInTest !== 0) {
+        isTestDeleted = res?.sectionInTest?.map((e) => {
+          return e.test.deleted
+        })
+      }
+      if (res?._count.sectionInTest === 0 || isTestDeleted?.includes(true)) {
+        isSectionDelete = true
+      } else {
+        let title = statusCheck.testDependentWarning
         deleteHandle = json<ActionData>(
           { errors: { title, status: 400, check: new Date() } },
           { status: 400 }
         )
-      })
+      }
+    })
+
+    if (isSectionDelete) {
+      await deleteSectionById(formData.get('id') as string)
+        .then((res) => {
+          deleteHandle = json<ActionData>(
+            { resp: { status: statusCheck.deletedSuccess } },
+
+            { status: 200 }
+          )
+        })
+        .catch((err) => {
+          let title = statusCheck.commonError
+          deleteHandle = json<ActionData>(
+            { errors: { title, status: 400, check: new Date() } },
+            { status: 400 }
+          )
+        })
+    }
     return deleteHandle
   }
 }
@@ -184,12 +206,12 @@ export default function SectionPage() {
       navigate(routes.sections, {
         replace: true,
       })
-    }else{
+    } else {
       navigate(`${routes.sections}/${selectedSection}${data?.filters}`, {
         replace: true,
       })
     }
-  }, [navigate,selectedSection])
+  }, [navigate, selectedSection])
 
   useEffect(() => {
     if (data.sections.length) {
@@ -229,7 +251,7 @@ export default function SectionPage() {
         })
       }
     }
-  }, [sectionActionData,data.selectedSectionId])
+  }, [sectionActionData, data.selectedSectionId])
 
   return (
     <AdminLayout>
